Trait node_runtime::StorageValue

pub trait StorageValue<T> where
    T: FullCodec, {
    type Query;
    fn hashed_key() -> [u8; 16];

    fn exists() -> bool;

    fn get() -> Self::Query;

    fn put<Arg>(val: Arg)
    where
        Arg: EncodeLike<T>;

    fn mutate<R, F>(f: F) -> R
    where
        F: FnOnce(&mut Self::Query) -> R;

    fn kill();

    fn take() -> Self::Query;

    fn append<Items, Item, EncodeLikeItem>(
        items: Items
    ) -> Result<(), &'static str>
    where
        EncodeLikeItem: EncodeLike<Item>,
        Item: Encode,
        Items: IntoIterator<Item = EncodeLikeItem>,
        T: EncodeAppend<Item = Item>,
        <Items as IntoIterator>::IntoIter: ExactSizeIterator;

    fn append_or_put<Items, Item, EncodeLikeItem>(items: Items)
    where
        EncodeLikeItem: EncodeLike<Item>,
        Item: Encode,
        Items: IntoIterator<Item = EncodeLikeItem> + Clone + EncodeLike<T>,
        T: EncodeAppend<Item = Item>,
        <Items as IntoIterator>::IntoIter: ExactSizeIterator;

    fn decode_len() -> Result<usize, &'static str>
    where
        T: DecodeLength + Len;
}

A trait for working with macro-generated storage values under the substrate storage API.

Associated Types

type Query

The type that get/take return.

Loading content...
Required methods

fn hashed_key() -> [u8; 16]

Get the storage key.

fn exists() -> bool

Does the value (explicitly) exist in storage?

fn get() -> Self::Query

Load the value from the provided storage instance.

fn put<Arg>(val: Arg) where
Arg: EncodeLike<T>,

Store a value under this key into the provided storage instance.

fn mutate<R, F>(f: F) -> R where
F: FnOnce(&mut Self::Query) -> R,

Mutate the value

fn kill()

Clear the storage value.

fn take() -> Self::Query

Take a value from storage, removing it afterwards.

fn append<Items, Item, EncodeLikeItem>(items: Items) -> Result<(), &'static str> where
EncodeLikeItem: EncodeLike<Item>,
Item: Encode,
Items: IntoIterator<Item = EncodeLikeItem>,
T: EncodeAppend<Item = Item>,
<Items as IntoIterator>::IntoIter: ExactSizeIterator,

Append the given item to the value in the storage.

T is required to implement codec::EncodeAppend.

fn append_or_put<Items, Item, EncodeLikeItem>(items: Items) where
EncodeLikeItem: EncodeLike<Item>,
Item: Encode,
Items: IntoIterator<Item = EncodeLikeItem> + Clone + EncodeLike<T>,
T: EncodeAppend<Item = Item>,
<Items as IntoIterator>::IntoIter: ExactSizeIterator,

Append the given items to the value in the storage.

T is required to implement Codec::EncodeAppend.

Upon any failure, it replaces items as the new value (assuming that the previous stored data is simply corrupt and no longer usable).

WARNING

use with care; if your use-case is not exactly as what this function is doing, you should use append and sensibly handle failure within the runtime code if it happens.

fn decode_len() -> Result<usize, &'static str> where
T: DecodeLength + Len,

Read the length of the value in a fast way, without decoding the entire value.

T is required to implement Codec::DecodeLength.

Loading content...