Crate node_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants
A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

BlockHashCount

EpochDuration

ExistentialDeposit

ExpectedBlockTime

GenesisConfig

MaxLocks

MinimumMultiplier

MinimumPeriod

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PointsLimit

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

TargetBlockFullness

TotalBandwidth

TransactionByteFee

Version

WeightLimit

Enums

Call

Event

OriginCaller

Constants

VERSION
This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap
Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address
The address format for describing accounts.

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Context
The type used as a helper for interpreting the sender of transactions.

Datalog

DigitalTwin

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Launch

RWS

RandomnessCollectiveFlip

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.