Crate node_runtime

The Substrate runtime. This can be compiled with ``#[no_std]`, ready for Wasm.

Modules

api

Structs

CurrencyToVoteHandler

GenesisConfig

Log
Wrapper for all possible log entries for the $trait runtime. Provides binary-compatible Encode/Decode implementations with the corresponding generic::DigestItem.

Perbill
Perbill is parts-per-billion. It stores a value between 0 and 1 in fixed point and provides a means to multiply some other value by that.

Permill
Permill is parts-per-million (i.e. after multiplying by this, divide by 1000000).

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

Enums

BalancesCall

Call

ConsensusCall

Event

InternalLog
All possible log entries for the $trait runtime. Encode/Decode implementations are auto-generated => it is not binary-compatible with generic::DigestItem.

Origin

StakerStatus
Indicates the initial status of the staker.

TimestampCall

Constants

VERSION
Runtime version.

Traits

BuildStorage
Complex storage builder stuff.

StorageValue
A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

Aura

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Consensus

ConsensusConfig

Contract

ContractConfig

Council

CouncilMotions

CouncilSeatsConfig

CouncilVoting

CouncilVotingConfig

Democracy

DemocracyConfig

Executive
Executive: handles dispatch to the various modules.

FinalityTracker

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Indices

IndicesConfig

Session

SessionConfig

SignedBlock
A Block signed with a Justification

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TimestampConfig

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.